CI base image: Remove the multi stage build
Maybe I'm doing a mistake because I miss the motivation for having a multi
step build in the first place: I don't see where files are generated in one
step to be erased in a following one. In any case, generated images are
actually smaller (!) than when / is recursively copied in a fresh image.

And we take advantage of having the problematic environment variables unset
(aka the compilation of 'opam-core' is fixed).